refactor: OptionChipBase 삭제
